Fix warning in GG_String::resize(unsigned int)
Avoids the GCC and Clang warning:

gg_string.cpp: In member function ‘void GG_String::resize(unsigned int)’:
gg_string.cpp:52:19: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

+++ src/gamegui/gg_string.h
@@ -102,7 +102,7 @@ class GG_String
       return m_capacity;
     }
 
-    inline int length(void)
+    inline size_t length(void)
     {
       return strlen(m_str);
     }
